gwl: Extract app group scroll box to its own module
